Разобралась с багом сервиса
