+++ src/model/destinations-model.js
@@ -1,7 +1,7 @@
 export default class DestinationsModel {
   constructor(service) {
     this.service = service;
-    this.destinations = service.getDestinations();
+    this.destinations = this.service.getDestinations();
   }
 
   get() {
